Move MockNetworkSession to new file.
